+++ cypress/utils/apps/query_enhancements/shared.js
@@ -103,3 +103,74 @@ export const getDefaultQuery = (datasetName, language) => {
       return `SELECT * FROM ${datasetName} LIMIT 10`;
   }
 };
+
+/**
+ * Gets Date and returns string of date in format MMM dd, yyyy @ HH:mm:ss.SSS
+ * @param {Date} date - date to format
+ * @returns {string} - in the format e.g. Jan 24, 2025 @ 16:20:08.000
+ * @example formatDate(new Date('2011-10-05T14:48:00.000Z'))
+ * // returns 'Oct 5, 2011 @ 14:48:00.000'
+ */
+export const formatDate = (date) => {
+  const months = [
+    'Jan',
+    'Feb',
+    'Mar',
+    'Apr',
+    'May',
+    'Jun',
+    'Jul',
+    'Aug',
+    'Sep',
+    'Oct',
+    'Nov',
+    'Dec',
+  ];
+
+  const month = months[date.getMonth()];
+  const day = date.getDate().toString();
+  const year = date.getFullYear();
+  const hours = date.getHours().toString().padStart(2, '0');
+  const minutes = date.getMinutes().toString().padStart(2, '0');
+  const seconds = date.getSeconds().toString().padStart(2, '0');
+  const milliseconds = date.getMilliseconds().toString().padStart(3, '0');
+
+  return `${month} ${day}, ${year} @ ${hours}:${minutes}:${seconds}.${milliseconds}`;
+};
+
+/**
+ * Adding thousand separators to numbers using locale settings
+ * @param {number} number - number value to format
+ * @returns {string}
+ * @example formatNumber(1000)
+ * // returns '1,000'
+ */
+const formatNumber = (number) => {
+  return number.toLocaleString();
+};
+
+/**
+ *  Formats values for UI display by:
+ * - Adding thousand separators to numbers using locale settings
+ * - Converting SQL/PPL datetime strings (yyyy-MM-dd HH:mm:ss.fff) to formatted dates (MMM dd, yyyy @ HH:mm:ss.fff)
+ * - Passing through other values unchanged
+ * @param {string|number} value - The value to format
+ * @returns {string} The formatted value
+ * @example
+ * formatValue(1000) // "1,000"
+ * formatValue("2025-01-24 16:20:08.000") // "Jan 24, 2025 @ 16:20:08.000"
+ */
+export const formatValue = (value) => {
+  switch (typeof value) {
+    case 'number':
+      return formatNumber(value);
+    case 'string':
+      // SQL and PPL date string uses format YYYY-MM-DD HH:MM:SS.SSS, we want to convert to ISO date.
+      const potentialDate = new Date(value.replace(' ', 'T'));
+      if (potentialDate instanceof Date && !isNaN(potentialDate)) {
+        return formatDate(potentialDate);
+      }
+    default:
+      return value;
+  }
+};
